requirements.txt
python-dotenv
anthropic